backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]

[tool.setuptools.package-dir]
backend = "app"

[project]
name = "workflow-automation"
version = "0.1.0"
description = "Workflow automation via serial number"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn[standard]==0.40.0",
    "azure-identity==1.25.1",
    "azure-cosmos==4.14.5",
    "azure-storage-blob==12.28.0",
    "azure-keyvault-secrets==4.10.0",
    "agent-framework==1.0.0b260130",
    "azure-ai-inference==1.0.0b9",
    "azure-ai-projects==2.0.0b3",
    "azure-ai-evaluation==1.14.0",
    "openai==2.16.0",
    "tiktoken==0.12.0",
    "fastmcp==3.1.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "azure-appconfiguration-provider>=2.1.0",
    "python-dotenv==1.2.1",
    "dependency-injector==4.48.3",
    "tenacity==9.1.2",
    "azure-monitor-opentelemetry==1.8.5",
    "azure-monitor-opentelemetry-exporter==1.0.0b46",
    "opentelemetry-api==1.39.0",
    "opentelemetry-sdk==1.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.1",
    "black==26.1.0",
    "ruff==0.14.14",
    "mypy==1.19.1",
    "types-requests==2.32.4.20260107",
    "debugpy==1.8.20",
]

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs | \.git | \.hg | \.mypy_cache | \.tox | \.venv | build | dist
)/
'''

[tool.ruff]
line-length = 120
target-version = "py312"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[[tool.mypy.overrides]]
module = "azure.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*", "*/site-packages/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
]
